Breaking: Start gutting component to rebuild it.
