+++ src/components/FileBrowser/FileBrowser.jsx
@@ -1,71 +1,46 @@
-import { useEffect, useState, useRef, useLayoutEffect, useCallback } from "react";
+import {
+    forwardRef,
+    useEffect,
+    useReducer,
+    useState,
+    useRef,
+    useLayoutEffect,
+    useMemo,
+    useCallback,
+    useImperativeHandle
+} from "react";
 import "./FileBrowser.scss";
 
-import { TreeNode } from "./TreeNode/TreeNode";
+import { Tree } from "./Tree/Tree";
 
 import { FileBrowserContext } from "./FileBrowserContext";
 
-import { setDefaultCollapsed, collapseTree, selectNode, flattenTree } from "./helper";
+import { fileBrowserReducer, initialState } from "./FileBrowserReducer";
 
-/**
- * Used for creating a stable callback function.
- */
-function useEvent(fn) {
-    const fnRef = useRef(fn);
-
-    useLayoutEffect(() => {
-        fnRef.current = fn;
-    });
-
-    return useCallback((...args) => {
-        return fnRef.current(...args);
-    }, []);
-}
+import {
+    setDefaultCollapsed,
+    collapseTree,
+    selectNode,
+    flattenTree,
+} from "./helper";
 
 /**
  * Renders a file browser.
- * 
+ *
  * @return {JSX}
  */
-export const FileBrowser = ({tree, onNodeSelect}) => {
-
-    const [nodes, setNodes] = useState([]);
-    const treeRef = useRef();
-
-    const handleFileClick = useEvent((node) => {
-        selectNode(treeRef.current, node);
-        drawTree();
-        if (onNodeSelect) {
-            onNodeSelect(node);
-        }
-    });
-
-    useEffect(() => {
-        treeRef.current = flattenTree(tree);
-        setDefaultCollapsed(treeRef.current);
-        drawTree();
-    }, []);
-
-    /**
-     * Draws the tree given the nodes and the collapsed state of each node.
-     */
-    const drawTree = () => {
-        const nodes = collapseTree(treeRef.current);
-        const rows = [];
-        nodes.forEach((node) => {
-            rows.push(<TreeNode key={node.id} node={node} id={node.name} onRowClick={handleFileClick}/>);
-        });
-        setNodes(rows);
-    }
+export const FileBrowser = forwardRef(({ }, ref) => {
+    const [state, dispatch] = useReducer(fileBrowserReducer, initialState);
 
     const addFileTree = useCallback((tree) => {
+        console.log("Adding file tree");
         dispatch({ type: "ADD_FILE_TREE", payload: tree });
     }, []);
 
     const api = useMemo(() => {
         return {
             state,
-            addFileTree
+            addFileTree,
         };
     }, [state, addFileTree]);
 
@@ -74,8 +49,8 @@ export const FileBrowser = ({tree, onNodeSelect}) => {
     return (
         <FileBrowserContext.Provider value={api}>
             <div className="file-browser">
-                {nodes}
+                <Tree />
             </div>
         </FileBrowserContext.Provider>
-    )
-}
+    );
+});
